+++ .gitignore
@@ -10,6 +10,7 @@ coverage
 
 # next.js
 .next/
+**/next-env.d.ts
 out/
 build
