+++ examples/appdir/tsconfig.json
@@ -1,7 +1,11 @@
 {
   "extends": "@uploadthing/tsconfig/nextjs.json",
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
-
-  "compilerOptions": { "plugins": [{ "name": "next" }] },
-  "exclude": ["node_modules"]
+  "compilerOptions": {
+    "plugins": [{ "name": "next" }],
+    "baseUrl": ".",
+    "paths": {
+      "~/*": ["./src/*"]
+    }
+  },
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"]
 }
