+++ examples/pagedir/package.json
@@ -14,7 +14,7 @@
     "@tanstack/react-query": "^4.28.0",
     "@uploadthing/react": "3.0.5",
     "autoprefixer": "10.4.14",
-    "next": "13.4.3",
+    "next": "13.4.4",
     "next-auth": "^4.22.1",
     "postcss": "8.4.22",
     "react": "18.2.0",
