+++ packages/config/eslint/package.json
