+++ packages/config/tsconfig/README.md
