+++ packages/config/tsconfig/nextjs.json
