+++ packages/config/tsconfig/react-library.json
