+++ packages/config/tsup/index.d.ts
@@ -1,0 +1,4 @@
+declare module "@uploadthing/tsup-config" {
+  import { Options } from "tsup";
+  export const config: Options;
+}
