+++ packages/config/tsup/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@uploadthing/tsup-config",
+  "version": "0.1.0",
+  "private": true,
+  "main": "index.js",
+  "types": "index.d.ts",
+  "devDependencies": {
+    "@types/node": "18.16.0",
+    "tsup": "6.7.0"
+  }
+}
