+++ packages/react/package.json
@@ -44,6 +44,8 @@
     "@types/react": "18.2.0",
     "@types/react-dom": "18.0.11",
     "@uploadthing/eslint-config": "0.1.0",
+    "@uploadthing/tsconfig": "0.1.0",
+    "@uploadthing/tsup-config": "0.1.0",
     "autoprefixer": "10.4.14",
     "eslint": "^8.40.0",
     "postcss": "8.4.22",
