+++ packages/uploadthing/package.json
@@ -45,8 +45,10 @@
   "devDependencies": {
     "@types/node": "18.16.0",
     "@uploadthing/eslint-config": "0.1.0",
+    "@uploadthing/tsconfig": "0.1.0",
+    "@uploadthing/tsup-config": "0.1.0",
     "eslint": "^8.40.0",
-    "next": "13.4.3",
+    "next": "13.4.4",
     "tsup": "6.7.0",
     "typescript": "5.1.0-beta",
     "vitest": "^0.30.1"
